add selectors for [diaryListReducer, myPageReducer] and apply ito mapStateToProps
